refactor of encoder side of laurium workflow
