+++ src/laurium/encoder_models/setfit/__init__.py
@@ -1,0 +1,1 @@
+"""SetFit-based architecture encoder models."""
