+++ src/laurium/encoder_models/transformers/__init__.py
@@ -1,0 +1,1 @@
+"""Transformers-based architecture encoder models."""
